planpalapp/requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.3.1
certifi==2025.7.14
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==45.0.5
daphne==4.2.1
dj-database-url==2.2.0
Django==5.2.6
django-celery-beat==2.8.1
django-ckeditor==6.7.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-js-asset==3.1.2
django-oauth-toolkit==3.0.1
django-stubs==5.2.2
django-stubs-ext==5.2.2
django-timezone-field==7.1
djangorestframework==3.14.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
firebase-admin==6.5.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jwcrypto==1.5.6
kombu==5.5.4
msgpack==1.1.1
mypy==1.17.1
mypy_extensions==1.1.0
mysqlclient==2.2.7
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
Pillow==10.0.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==4.6.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
types-cffi==1.17.0.20250822
types-pyOpenSSL==24.1.0.20240722
types-PyYAML==6.0.12.20250822
types-redis==4.6.0.20241004
types-requests==2.32.4.20250809
types-setuptools==80.9.0.20250822
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
zope.interface==8.0
